+++ components/tabs/PropsType.tsx
@@ -1,9 +1,10 @@
 export default interface PropsType {
   value?: number;
   defaultValue?: number;
-  scrollThreshold: number;
   lineWidth?: string | number;
   disabled?: boolean;
-  canSwipe?: boolean;
+  swipeable?: boolean;
+  scrollable?: boolean;
+  direction?: 'horizontal' | 'vertical';
   onChange?: (index?: number) => void;
 }