+++ components/tabs/__tests__/index.test.jsx
@@ -18,9 +18,23 @@ describe('Tab', () => {
     expect(toJson(wrapper)).toMatchSnapshot();
   });
 
+  it('renders correctly vertical', () => {
+    const wrapper = render(
+      <Tabs direction="vertical">
+        <Tabs.Panel title="选项卡1">
+          <div>试试点我左滑</div>
+        </Tabs.Panel>
+        <Tabs.Panel title="选项卡2">
+          <div>试试点我右滑</div>
+        </Tabs.Panel>
+      </Tabs>,
+    );
+    expect(toJson(wrapper)).toMatchSnapshot();
+  });
+
   it('canSwipe', () => {
     const wrapper = render(
-      <Tabs canSwipe>
+      <Tabs swipeable>
         <Tabs.Panel title="选项卡1">
           <div>试试点我左滑</div>
         </Tabs.Panel>
@@ -87,7 +101,7 @@ describe('Tab', () => {
   it('scroll tabs', () => {
     const onChange = jest.fn();
     const wrapper = mount(
-      <Tabs canSwipe onChange={onChange} scrollThreshold={3}>
+      <Tabs swipeable onChange={onChange} scrollable>
         <Tabs.Panel title="选项卡1" disabled>
           <div>试试点我左滑</div>
         </Tabs.Panel>