+++ components/tabs/util/index.ts
@@ -1,7 +1,3 @@
-import { raf, cancelRaf } from '../../utils/raf';
-
-let scrollLeftRafId: number;
-
 export function getTransformPropValue(v: any) {
   return {
     transform: v,
@@ -13,26 +9,4 @@ export function getTransformPropValue(v: any) {
 export function getPxStyle(value: number | string, unit = 'px', vertical = false) {
   value = vertical ? `0px, ${value}${unit}, 0px` : `${value}${unit}, 0px, 0px`;
   return `translate3d(${value})`;
-}
-
-
-export function scrollLeftTo(
-  scroller: HTMLElement,
-  to: number,
-  duration: number,
-) {
-  cancelRaf(scrollLeftRafId);
-
-  let count = 0;
-  const from = scroller.scrollLeft;
-  const frames = duration === 0 ? 1 : Math.round((duration * 1000) / 16);
-  function animate() {
-    scroller.scrollLeft += (to - from) / frames;
-    count += 1;
-    if (count < frames) {
-      scrollLeftRafId = raf(animate);
-    }
-  }
-
-  animate();
 }