+++ examples-rn/babel.config.js
@@ -1,1 +1,0 @@
-module.exports = require('@zarm-design/cli/lib/config/babelConfig/native');